\name{CorreccionPorOtrasVariables}
\alias{CorreccionPorOtrasVariables}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{CorreccionPorOtrasVariables}
\description{Calcula el factor de correccion por otras variables}
\usage{
CorreccionPorOtrasVariables(VariablesDeLaMuestra, CasoEstimacion, Parametros)
}
\arguments{
  \item{VariablesDeLaMuestra}{Matriz de variables de la muestra}
  \item{CasoEstimacion}{Caso a estimar}
  \item{Parametros}{Parametros de calculo del modelo}
}
\details{}
\value{Devuelve los valores de correccion para cada muestra}
\references{}
\author{}
\note{}
\examples{}
\keyword{}
\keyword{}
