\name{LeerArchivos}
\alias{LeerArchivos}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Leer archivos con los datos para estimar la muestra
}
\description{Lee los archivos CasoEstimacion.csv, ParametrosCalculo.csv y DatosMuestra.csv en la ubicacion indicada.}
\usage{
LeerArchivos(Directorio)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Directorio}{Directorio donde estan ubicados los archivos}
}
\details{}
\value{}
\references{}
\author{}
\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{}
\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{}
\keyword{}% __ONLY ONE__ keyword per line
