\name{ObtieneDatosDelModelo}
\alias{ObtieneDatosDelModelo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Obtiene los coeficientes del modelo a partir de la tipologia a estimar y los parametros}
\description{Obtiene los coeficientes del modelo a partir de la tipologia a estimar y los parametros}
\usage{
ObtieneDatosDelModelo(ParametrosDelModelo, Tipologia)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ParametrosDelModelo}{Parametros del modelo para los calculos}
  \item{Tipologia}{Tipologia del caso a estimar}
}
\details{}
\value{}
\references{}
\author{}
\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{}
\examples{}
\keyword{}
\keyword{}
