\name{SeleccionarOtrasVariables}
\alias{SeleccionarOtrasVariables}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Selecciona las variables NumeroDeAscensores, ExteriorInterior, NumeroDePlanta, y DatosDeLaPlanta de la muestra recibida}
\description{}
\usage{
SeleccionarOtrasVariables(Muestra)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Muestra}{Muestra de la que extraer las columnas NumeroDeAscensores, ExteriorInterior, NumeroDePlanta, y DatosDeLaPlanta
}
}
\details{}
\value{}
\references{}
\author{}
\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{}
\examples{}
\keyword{}
\keyword{}% __ONLY ONE__ keyword per line
