\name{Tasacion-package}
\alias{Tasacion-package}
\alias{Tasacion}
\docType{package}
\title{Genera una tasacion}
\description{Genera una tasacion aleatoria}
\details{
\tabular{ll}{
Package: \tab Tasacion\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-16\cr
License: \tab ---\cr
}
Genera una tasacion
}
\author{Federico Munyoz

Maintainer: Federico Munyoz <federico.munyoz@innocv.com>
}
\references{}

\keyword{ package }
\seealso{}
\examples{
Tasacion("c:/tmp")
}
